flask-socketio/requirements.txt
Flask==2.1.1
Werkzeug==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.0
itsdangerous==2.1.1
click==8.1.2
gunicorn==20.1.0
openai
flask-socketio
